requirements.txt
Flask==2.3.2
gunicorn==21.2.0
requests==2.31.0
validators==0.35.0
python-dotenv==1.0.0
psycopg[binary]
setuptools==80.9.0
beautifulsoup4==4.12.3